Refactor composeLineWrapAPI
Pull the toggle logic out of the API extension, only exposing get/set
for an editor's wrap width and the default wrap width. This may make it
easier to eventually sell it to TB.

The default wrap width is also returned as 0 if the editor is in HTML
mode. While compose.getComposeDetails suggests to provide that
information, it crashes when called to early in TB 78.4.0. So we stick
with the API extension here as well.

Signed-off-by: Jan Kiszka <jan.kiszka@web.de>