Fix application of line_wrap default on new editor windows
It's too early to call ComposeLineWrap.setEditorWrapWidth from the
onCreated handler, gMsgCompose.editor is still undefined at this point.
Doing that from initial onFocusChanged call is fine, though.

So migrate the setup into the onFocusChanged handler, using a set of the
windows IDs to track the initialization state of windows.

Signed-off-by: Jan Kiszka <jan.kiszka@web.de>