Wait for editor reference in gMsgCompose
Even onFocusChanged events may come too early to have the editor
reference in gMsgCompose been initialized. That will cause the initial
setWrapWidth to fail with an exception. Seen first with TB 91.8.1 under
Windows. It might have happened before, just more rarely.

This add a delayed polling loop to getMsgCompose that will check for
editor being non-null every 100 ms. With that in place, the check for
this in getWrapWidth can also be removed.

Signed-off-by: Jan Kiszka <jan.kiszka@web.de>