Add support for detecting patches
Patches require line wrapping to be turned off - one of the main reasons
why this extension exists. Improve this use case by detecting patches
using the commonly used tag "[PATCH]" (or also "[PATCH ...]"). If found
while setting up a composer window, turn off line wrapping directly. If
found before a message is to be send and line wrapping is enabled, give
the user a warning permit to cancel the submission, giving a chance to
fix the situation.

To account for potential conflicts of this feature, make it optional
while enabling it by default.

Signed-off-by: Jan Kiszka <jan.kiszka@web.de>